requirements.txt
slack-sdk==3.23.0
anthropic==0.25.0
python-dotenv==1.0.0
supabase==2.7.4
httpx==0.25.2
websockets==12.0
